LayoutTests/crypto/workers/subtle/aes-indexeddb-expected.txt

[Worker] Test generating an AES key using AES-CBC algorithm in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/aes-indexeddb.js
[Worker] Importing a key...
PASS [Worker] retrievedKey.type is 'secret'
PASS [Worker] retrievedKey.extractable is true
PASS [Worker] retrievedKey.algorithm.name is 'AES-CBC'
PASS [Worker] retrievedKey.algorithm.length is 128
PASS [Worker] retrievedKey.usages is ['decrypt', 'encrypt', 'unwrapKey', 'wrapKey']
PASS successfullyParsed is true

TEST COMPLETE
